+++ MotherCore/Modules/IModule.cs
@@ -1,23 +1,6 @@
 ﻿using Sandbox.Game.EntityComponents;
 using Sandbox.ModAPI.Ingame;
-using Sandbox.ModAPI.Interfaces;
-using SpaceEngineers.Game.ModAPI.Ingame;
-using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Linq;
-using System.Text;
-using VRage;
-using VRage.Collections;
-using VRage.Game;
-using VRage.Game.Components;
-using VRage.Game.GUI.TextPanel;
-using VRage.Game.ModAPI.Ingame;
-using VRage.Game.ModAPI.Ingame.Utilities;
-using VRage.Game.ObjectBuilders.Definitions;
-using VRage.Scripting.MemorySafeTypes;
-using VRageMath;
 
 namespace IngameScript
 {
@@ -27,7 +10,6 @@ namespace IngameScript
     /// </summary>
     public interface IModule
     {
-   
         /// <summary>
         /// Boot the module. This is called before the module is run for the first 
         /// time and is the ideal method to define dependencies on other modules.